requirements.txt
Django==5.2
pandas==2.2.3
numpy==2.2.4
openpyxl==3.1.5
python-dateutil==2.9.0
pytz==2025.2
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-environ==0.11.2
gunicorn==21.2.0
whitenoise==6.6.0 